feat: [OBI-150] Added change password logic
